# contracts/opportunity.openapi.yaml
openapi: 3.0.3
info:
  title: crm-opportunity · 商机管理
  version: 1.0.0
  description: |
    对外 REST 面。内部 gRPC，外部 HTTP（§2.1）。
    ⚠️ 不暴露 BatchGetOpportunities——组件间协议，给 BFF 层 GraphQL
    Resolver 防 N+1 用（设计计划 §3、§3.8）。
    ⚠️ 赢单单独一个权限键 crm.opportunity.win，不与普通编辑合并——它会
    触发建单、锁库存、生成应收一整条链路，是本组件唯一有下游财务影响的
    动作（设计计划 §3）。

servers:
  - url: /crm/opportunity

# ⚠️ /healthz 不在这份契约里：be-sdk-go 的 NewGinEngine 给每个组件统一挂的，
# 路径固定是裸根路径 /healthz（§12.3.6）。

paths:
  /opportunities:
    get:
      summary: ListOpportunities（漏斗视图的数据源）
      operationId: listOpportunities
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: 游标分页（决策 53），无 offset 参数——深分页在契约层面不可表达
        - name: page_size
          in: query
          schema: { type: integer }
        - name: customer_id
          in: query
          schema: { type: string }
        - name: status_filter
          in: query
          schema: { $ref: "#/components/schemas/OpportunityStatus" }
        - name: stage_id_filter
          in: query
          schema: { type: string }
        - name: view
          in: query
          schema: { type: string, enum: [mine, dept] }
          description: "省略默认 dept（本部门及下级）；mine = 我的商机（owner_id 相等）"
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
          description: 不传时间范围时框架层自动注入最近 90 天（§11.4.1）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunities:
                    type: array
                    items: { $ref: "#/components/schemas/Opportunity" }
                  next_cursor: { type: string }
    post:
      summary: CreateOpportunity（建档，校验客户与产品）
      operationId: createOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOpportunityRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }

  /opportunities/{id}:
    get:
      summary: GetOpportunity
      operationId: getOpportunity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }
    patch:
      summary: UpdateOpportunity（改金额/预计成交日/负责人）
      operationId: updateOpportunity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOpportunityRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }

  /opportunities/{id}/stage:
    post:
      summary: ChangeStage（阶段推进，写历史）
      operationId: changeOpportunityStage
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeStageRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }

  /opportunities/{id}/win:
    post:
      summary: MarkWon（⭐ 链路的扳机：转 WON + 发 crm.opportunity.won.v1）
      operationId: markOpportunityWon
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IdempotentRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }

  /opportunities/{id}/lose:
    post:
      summary: MarkLost（转 LOST，要求填原因）
      operationId: markOpportunityLost
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkLostRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Opportunity" }

  /stages:
    get:
      summary: ListStages（阶段是数据不是代码，迁移播种，供前端阶段选择器/漏斗视图读）
      operationId: listOpportunityStages
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  stages:
                    type: array
                    items: { $ref: "#/components/schemas/OpportunityStage" }

components:
  schemas:
    OpportunityStatus:
      type: string
      enum: [OPEN, WON, LOST, CANCELLED]

    OpportunityStage:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        sort_order: { type: integer }
        default_probability: { type: integer }
        is_final: { type: boolean }

    OpportunityItem:
      type: object
      properties:
        product_id: { type: string }
        product_sku: { type: string }
        product_name: { type: string }
        uom_id: { type: string }
        qty: { type: string, description: decimal-as-string }
        quoted_unit_price:
          type: string
          description: "⚠️ 报价快照供参考，不是权威价目——erp-sales 赢单转订单时会用自己的定价引擎重算（设计计划 §2.1）"
        subtotal:
          type: string
          description: 服务端算好直接给，不让调用方自己算

    Opportunity:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        customer_id: { type: string }
        customer_name: { type: string }
        owner_id: { type: string }
        dept_id: { type: string }
        dept_path: { type: string }
        stage_id: { type: string }
        stage_name: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OpportunityItem" }
        expected_amount: { type: string, description: 金额一律用 string 传 decimal，绝不用 double }
        probability: { type: integer, description: "0-100，阶段带默认值 + 允许手工覆盖" }
        weighted_amount: { type: string, description: "⭐ = expected_amount × probability，算出来不存" }
        expected_close_date: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/OpportunityStatus" }
        currency: { type: string }
        lost_reason: { type: string }
        won_at: { type: string, format: date-time }
        version: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    IdempotentRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key:
          type: string
          description: 跨重试幂等键（§4.6），被调用方用唯一约束去重

    CreateOpportunityItemInput:
      type: object
      required: [product_id, qty, quoted_unit_price]
      properties:
        product_id: { type: string }
        qty: { type: string }
        quoted_unit_price: { type: string }

    CreateOpportunityRequest:
      type: object
      required: [idempotency_key, name, customer_id, items, expected_amount]
      properties:
        idempotency_key: { type: string }
        name: { type: string }
        customer_id: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/CreateOpportunityItemInput" }
        expected_amount: { type: string }
        probability:
          type: integer
          description: 留空则用初始阶段的 default_probability
        expected_close_date: { type: string, format: date-time }
        currency: { type: string, description: 留空则服务端按本位币默认值处理 }

    UpdateOpportunityRequest:
      type: object
      required: [idempotency_key, version]
      properties:
        idempotency_key: { type: string }
        version: { type: integer, format: int64, description: 乐观锁：与库里不一致则拒绝 }
        name: { type: string }
        expected_amount: { type: string }
        expected_close_date: { type: string, format: date-time }
        owner_id: { type: string, description: 改负责人 }

    ChangeStageRequest:
      type: object
      required: [idempotency_key, to_stage_id]
      properties:
        idempotency_key: { type: string }
        to_stage_id: { type: string }
        probability_override:
          type: integer
          description: 留空则用目标阶段的 default_probability

    MarkLostRequest:
      type: object
      required: [idempotency_key, reason]
      properties:
        idempotency_key: { type: string }
        reason: { type: string, description: 必填：失单必须记原因 }